extracted DatesDataExtractor
